hyprland/workspaces: break up implementations
